+++ CHANGELOG.md
@@ -8,6 +8,8 @@
   - added button in overlay to logout deom portal
 - Dataspace
   - Updated dataspace provider links with quick links
+- Company Certificate
+  - Connect listing page, upload certificate and details page to back end api
 - Delete ownuser Redirect URL
   - fixed logout redirect url issue
 - Consent Pop-Up after registration